packages/backend/pyproject.toml
[project]
name = "yuntongbu-backend"
version = "0.1.0"
description = "FastAPI backend for synchronized mpv playback with Emby integration."
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.21.0",
  "cryptography>=45.0.0",
  "fastapi>=0.116.0",
  "httpx>=0.28.1",
  "orjson>=3.11.0",
  "passlib[bcrypt]>=1.7.4",
  "pydantic-settings>=2.10.1",
  "python-jose[cryptography]>=3.5.0",
  "sqlalchemy>=2.0.41",
  "uvicorn[standard]>=0.35.0",
  "yuntongbu-shared-protocol",
]

[project.scripts]
yuntongbu-backend = "yuntongbu_backend.main:main"

[tool.uv.sources]
yuntongbu-shared-protocol = { workspace = true }

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yuntongbu_backend"]
